# backend/internal/trainers/ports/http/openapi.gen.go
// Package http provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package http

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"backend/internal/trainers/domain"

	"github.com/labstack/echo/v5"
	"github.com/oapi-codegen/runtime"
)

// Defines values for GetTrainerHoursByUuidParamsStatus.
const (
	Availability      GetTrainerHoursByUuidParamsStatus = "availability"
	NotAvailability   GetTrainerHoursByUuidParamsStatus = "not-availability"
	TrainingScheduled GetTrainerHoursByUuidParamsStatus = "training-scheduled"
)

// Valid indicates whether the value is a known member of the GetTrainerHoursByUuidParamsStatus enum.
func (e GetTrainerHoursByUuidParamsStatus) Valid() bool {
	switch e {
	case Availability:
		return true
	case NotAvailability:
		return true
	case TrainingScheduled:
		return true
	default:
		return false
	}
}

// Date defines model for Date.
type Date struct {
	Date         time.Time `json:"date"`
	HasFreeHours bool      `json:"hasFreeHours"`
	Hours        []Hour    `json:"hours"`
}

// Error defines model for Error.
type Error struct {
	Details *[]ErrorDetail `json:"details,omitempty"`
	Message string         `json:"message"`
	Slug    string         `json:"slug"`
}

// ErrorDetail defines model for ErrorDetail.
type ErrorDetail struct {
	EntityId   *string `json:"entity_id,omitempty"`
	EntityType *string `json:"entity_type,omitempty"`
	ErrorSlug  *string `json:"error_slug,omitempty"`
	Message    *string `json:"message,omitempty"`
}

// Hour defines model for Hour.
type Hour struct {
	Hour   time.Time  `json:"hour"`
	Status HourStatus `json:"status"`
}

// HourStatus defines model for HourStatus.
type HourStatus = domain.HourStatus

// HoursRequest defines model for HoursRequest.
type HoursRequest struct {
	Hours []time.Time `json:"hours"`
}

// Conflict defines model for Conflict.
type Conflict = Error

// InvalidInput defines model for InvalidInput.
type InvalidInput = Error

// Unauthorized defines model for Unauthorized.
type Unauthorized = Error

// Hours defines model for Hours.
type Hours = HoursRequest

// GetTrainerHoursParams defines parameters for GetTrainerHours.
type GetTrainerHoursParams struct {
	From time.Time `form:"from" json:"from"`
	To   time.Time `form:"to" json:"to"`
}

// GetTrainerHoursByUuidParams defines parameters for GetTrainerHoursByUuid.
type GetTrainerHoursByUuidParams struct {
	From   time.Time                          `form:"from" json:"from"`
	To     time.Time                          `form:"to" json:"to"`
	Status *GetTrainerHoursByUuidParamsStatus `form:"status,omitempty" json:"status,omitempty"`
}

// GetTrainerHoursByUuidParamsStatus defines parameters for GetTrainerHoursByUuid.
type GetTrainerHoursByUuidParamsStatus string

// MakeHoursAvailableJSONRequestBody defines body for MakeHoursAvailable for application/json ContentType.
type MakeHoursAvailableJSONRequestBody = HoursRequest

// MakeHoursNotAvailableJSONRequestBody defines body for MakeHoursNotAvailable for application/json ContentType.
type MakeHoursNotAvailableJSONRequestBody = HoursRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// GetTrainerHours Get the trainer's hours in a date range
	// (GET /api/v1/trainer/hours)
	GetTrainerHours(ctx *echo.Context, params GetTrainerHoursParams) error
	// MakeHoursAvailable Make hours available
	// (PUT /api/v1/trainer/hours/available)
	MakeHoursAvailable(ctx *echo.Context) error
	// MakeHoursNotAvailable Make hours not available
	// (PUT /api/v1/trainer/hours/not-available)
	MakeHoursNotAvailable(ctx *echo.Context) error
	// GetTrainerHoursByUuid Get a trainer's hours in a date range
	// (GET /api/v1/trainers/{trainerUuid}/hours)
	GetTrainerHoursByUuid(ctx *echo.Context, trainerUuid string, params GetTrainerHoursByUuidParams) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// GetTrainerHours converts echo context to params.
func (w *ServerInterfaceWrapper) GetTrainerHours(ctx *echo.Context) error {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetTrainerHoursParams
	// ------------- Required query parameter "from" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "from", ctx.QueryParams(), &params.From, runtime.BindQueryParameterOptions{Type: "string", Format: "date-time"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter from: %s", err))
	}

	// ------------- Required query parameter "to" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "to", ctx.QueryParams(), &params.To, runtime.BindQueryParameterOptions{Type: "string", Format: "date-time"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter to: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetTrainerHours(ctx, params)
	return err
}

// MakeHoursAvailable converts echo context to params.
func (w *ServerInterfaceWrapper) MakeHoursAvailable(ctx *echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.MakeHoursAvailable(ctx)
	return err
}

// MakeHoursNotAvailable converts echo context to params.
func (w *ServerInterfaceWrapper) MakeHoursNotAvailable(ctx *echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.MakeHoursNotAvailable(ctx)
	return err
}

// GetTrainerHoursByUuid converts echo context to params.
func (w *ServerInterfaceWrapper) GetTrainerHoursByUuid(ctx *echo.Context) error {
	var err error
	// ------------- Path parameter "trainerUuid" -------------
	var trainerUuid string

	err = runtime.BindStyledParameterWithOptions("simple", "trainerUuid", ctx.Param("trainerUuid"), &trainerUuid, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: ctx.Request().URL.RawPath == ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter trainerUuid: %s", err))
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params GetTrainerHoursByUuidParams
	// ------------- Required query parameter "from" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "from", ctx.QueryParams(), &params.From, runtime.BindQueryParameterOptions{Type: "string", Format: "date-time"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter from: %s", err))
	}

	// ------------- Required query parameter "to" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "to", ctx.QueryParams(), &params.To, runtime.BindQueryParameterOptions{Type: "string", Format: "date-time"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter to: %s", err))
	}

	// ------------- Optional query parameter "status" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "status", ctx.QueryParams(), &params.Status, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter status: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetTrainerHoursByUuid(ctx, trainerUuid, params)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
}

// RegisterHandlersOptions configures RegisterHandlersWithOptions.
type RegisterHandlersOptions struct {
	// BaseURL is prepended to every registered path so the API can be served
	// under a prefix.
	BaseURL string
	// OperationMiddlewares lets the caller attach per-operation middleware at
	// registration time. The map key is the OpenAPI `operationId` value as it
	// appears in the spec (the raw, un-normalized form). Operations that have
	// no entry are registered with no extra middleware. A nil map disables
	// per-operation middleware entirely.
	OperationMiddlewares map[string][]echo.MiddlewareFunc
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{})
}

// RegisterHandlersWithBaseURL registers handlers and prepends BaseURL to the
// paths so the API can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{BaseURL: baseURL})
}

// RegisterHandlersWithOptions registers handlers using the supplied options,
// including any per-operation middleware.
func RegisterHandlersWithOptions(router EchoRouter, si ServerInterface, options RegisterHandlersOptions) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(options.BaseURL+"/api/v1/trainer/hours", wrapper.GetTrainerHours, options.OperationMiddlewares["getTrainerHours"]...)
	router.GET(options.BaseURL+"/api/v1/trainers/:trainerUuid/hours", wrapper.GetTrainerHoursByUuid, options.OperationMiddlewares["getTrainerHoursByUuid"]...)
	router.PUT(options.BaseURL+"/api/v1/trainer/hours/available", wrapper.MakeHoursAvailable, options.OperationMiddlewares["makeHoursAvailable"]...)
	router.PUT(options.BaseURL+"/api/v1/trainer/hours/not-available", wrapper.MakeHoursNotAvailable, options.OperationMiddlewares["makeHoursNotAvailable"]...)

}

type ConflictJSONResponse Error

type InvalidInputJSONResponse Error

type UnauthorizedJSONResponse Error

type GetTrainerHoursRequestObject struct {
	Params GetTrainerHoursParams
}

type GetTrainerHoursResponseObject interface {
	VisitGetTrainerHoursResponse(w http.ResponseWriter) error
}

type GetTrainerHours200JSONResponse []Date

func (response GetTrainerHours200JSONResponse) VisitGetTrainerHoursResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetTrainerHours400JSONResponse struct{ InvalidInputJSONResponse }

func (response GetTrainerHours400JSONResponse) VisitGetTrainerHoursResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type GetTrainerHours401JSONResponse struct{ UnauthorizedJSONResponse }

func (response GetTrainerHours401JSONResponse) VisitGetTrainerHoursResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

type MakeHoursAvailableRequestObject struct {
	Body *MakeHoursAvailableJSONRequestBody
}

type MakeHoursAvailableResponseObject interface {
	VisitMakeHoursAvailableResponse(w http.ResponseWriter) error
}

type MakeHoursAvailable204Response struct {
}

func (response MakeHoursAvailable204Response) VisitMakeHoursAvailableResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type MakeHoursAvailable400JSONResponse struct{ InvalidInputJSONResponse }

func (response MakeHoursAvailable400JSONResponse) VisitMakeHoursAvailableResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type MakeHoursAvailable401JSONResponse struct{ UnauthorizedJSONResponse }

func (response MakeHoursAvailable401JSONResponse) VisitMakeHoursAvailableResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

type MakeHoursAvailable409JSONResponse struct{ ConflictJSONResponse }

func (response MakeHoursAvailable409JSONResponse) VisitMakeHoursAvailableResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(409)
	_, err := buf.WriteTo(w)
	return err
}

type MakeHoursNotAvailableRequestObject struct {
	Body *MakeHoursNotAvailableJSONRequestBody
}

type MakeHoursNotAvailableResponseObject interface {
	VisitMakeHoursNotAvailableResponse(w http.ResponseWriter) error
}

type MakeHoursNotAvailable204Response struct {
}

func (response MakeHoursNotAvailable204Response) VisitMakeHoursNotAvailableResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type MakeHoursNotAvailable400JSONResponse struct{ InvalidInputJSONResponse }

func (response MakeHoursNotAvailable400JSONResponse) VisitMakeHoursNotAvailableResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type MakeHoursNotAvailable401JSONResponse struct{ UnauthorizedJSONResponse }

func (response MakeHoursNotAvailable401JSONResponse) VisitMakeHoursNotAvailableResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

type MakeHoursNotAvailable409JSONResponse struct{ ConflictJSONResponse }

func (response MakeHoursNotAvailable409JSONResponse) VisitMakeHoursNotAvailableResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(409)
	_, err := buf.WriteTo(w)
	return err
}

type GetTrainerHoursByUuidRequestObject struct {
	TrainerUuid string `json:"trainerUuid"`
	Params      GetTrainerHoursByUuidParams
}

type GetTrainerHoursByUuidResponseObject interface {
	VisitGetTrainerHoursByUuidResponse(w http.ResponseWriter) error
}

type GetTrainerHoursByUuid200JSONResponse []Date

func (response GetTrainerHoursByUuid200JSONResponse) VisitGetTrainerHoursByUuidResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetTrainerHoursByUuid400JSONResponse struct{ InvalidInputJSONResponse }

func (response GetTrainerHoursByUuid400JSONResponse) VisitGetTrainerHoursByUuidResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type GetTrainerHoursByUuid401JSONResponse struct{ UnauthorizedJSONResponse }

func (response GetTrainerHoursByUuid401JSONResponse) VisitGetTrainerHoursByUuidResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// GetTrainerHours Get the trainer's hours in a date range
	// (GET /api/v1/trainer/hours)
	GetTrainerHours(ctx context.Context, request GetTrainerHoursRequestObject) (GetTrainerHoursResponseObject, error)
	// MakeHoursAvailable Make hours available
	// (PUT /api/v1/trainer/hours/available)
	MakeHoursAvailable(ctx context.Context, request MakeHoursAvailableRequestObject) (MakeHoursAvailableResponseObject, error)
	// MakeHoursNotAvailable Make hours not available
	// (PUT /api/v1/trainer/hours/not-available)
	MakeHoursNotAvailable(ctx context.Context, request MakeHoursNotAvailableRequestObject) (MakeHoursNotAvailableResponseObject, error)
	// GetTrainerHoursByUuid Get a trainer's hours in a date range
	// (GET /api/v1/trainers/{trainerUuid}/hours)
	GetTrainerHoursByUuid(ctx context.Context, request GetTrainerHoursByUuidRequestObject) (GetTrainerHoursByUuidResponseObject, error)
}

type StrictHandlerFunc func(ctx *echo.Context, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// GetTrainerHours operation middleware
func (sh *strictHandler) GetTrainerHours(ctx *echo.Context, params GetTrainerHoursParams) error {
	var request GetTrainerHoursRequestObject

	request.Params = params

	handler := func(ctx *echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetTrainerHours(ctx.Request().Context(), request.(GetTrainerHoursRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetTrainerHours")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetTrainerHoursResponseObject); ok {
		return validResponse.VisitGetTrainerHoursResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// MakeHoursAvailable operation middleware
func (sh *strictHandler) MakeHoursAvailable(ctx *echo.Context) error {
	var request MakeHoursAvailableRequestObject

	var body MakeHoursAvailableJSONRequestBody
	var err error
	if _, ok := ctx.Echo().Binder.(*echo.DefaultBinder); ok {
		// Bind only the request body, so that path and query parameters
		// are not also bound into the body struct.
		err = echo.BindBody(ctx, &body)
	} else {
		// A custom binder is installed on the Echo instance; defer to it
		// entirely, since echo.Binder does not expose body-only binding.
		err = ctx.Bind(&body)
	}
	if err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx *echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.MakeHoursAvailable(ctx.Request().Context(), request.(MakeHoursAvailableRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "MakeHoursAvailable")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(MakeHoursAvailableResponseObject); ok {
		return validResponse.VisitMakeHoursAvailableResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// MakeHoursNotAvailable operation middleware
func (sh *strictHandler) MakeHoursNotAvailable(ctx *echo.Context) error {
	var request MakeHoursNotAvailableRequestObject

	var body MakeHoursNotAvailableJSONRequestBody
	var err error
	if _, ok := ctx.Echo().Binder.(*echo.DefaultBinder); ok {
		// Bind only the request body, so that path and query parameters
		// are not also bound into the body struct.
		err = echo.BindBody(ctx, &body)
	} else {
		// A custom binder is installed on the Echo instance; defer to it
		// entirely, since echo.Binder does not expose body-only binding.
		err = ctx.Bind(&body)
	}
	if err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx *echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.MakeHoursNotAvailable(ctx.Request().Context(), request.(MakeHoursNotAvailableRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "MakeHoursNotAvailable")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(MakeHoursNotAvailableResponseObject); ok {
		return validResponse.VisitMakeHoursNotAvailableResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// GetTrainerHoursByUuid operation middleware
func (sh *strictHandler) GetTrainerHoursByUuid(ctx *echo.Context, trainerUuid string, params GetTrainerHoursByUuidParams) error {
	var request GetTrainerHoursByUuidRequestObject

	request.TrainerUuid = trainerUuid
	request.Params = params

	handler := func(ctx *echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetTrainerHoursByUuid(ctx.Request().Context(), request.(GetTrainerHoursByUuidRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetTrainerHoursByUuid")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetTrainerHoursByUuidResponseObject); ok {
		return validResponse.VisitGetTrainerHoursByUuidResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}